multifile but typed stub doesn't work
