+++ core/src/main/java/io/temporal/samples/earlyreturn/TransactionActivities.java
@@ -1,0 +1,16 @@
+package io.temporal.samples.earlyreturn;
+
+import io.temporal.activity.ActivityInterface;
+import io.temporal.activity.ActivityMethod;
+
+@ActivityInterface
+public interface TransactionActivities {
+  @ActivityMethod
+  Transaction initTransaction(Transaction tx);
+
+  @ActivityMethod
+  void cancelTransaction(Transaction tx);
+
+  @ActivityMethod
+  void completeTransaction(Transaction tx);
+}
